Move each GraphQL configuration type into its own file
