+++ Sources/GraphQLCodegen/Configuration/Configuration.Input.swift
@@ -3,16 +3,6 @@ import Foundation
 extension Configuration {
     /// Options controlling how to ingest the GraphQL schema and GraphQL operations.
     public struct Input: Sendable {
-        /// Controls whether deprecated GraphQL schema members remain available to generated code.
-        public enum DeprecationPolicy: Sendable {
-            /// Include deprecated schema members. Generated declarations are annotated with a Swift warning where supported,
-            /// and code generation warns when an operation uses a deprecated argument.
-            case include
-
-            /// Exclude deprecated schema members and fail code generation when an operation uses one.
-            case exclude
-        }
-
         /// Call this function to create a new `Input` instance.
         ///
         /// - Parameters:
