+++ Sources/GraphQLCodegen/Configuration/Configuration.Output.Schema.Enums.CaseConversion.Case.swift
@@ -1,0 +1,6 @@
+extension Configuration.Output.Schema.Enums.CaseConversion {
+    public enum Case: Sendable {
+        case lowerCamel // thisIsCamelCase
+        case macro // THIS_IS_MACRO_CASE
+    }
+}
